Move error-handling to ak._v2._util.
