+++ src/awkward/_v2/contents/bitmaskedarray.py
@@ -6,7 +6,6 @@ import math
 
 import awkward as ak
 from awkward._v2.index import Index
-from awkward._v2._slicing import nested_indexerror
 from awkward._v2.contents.content import Content
 from awkward._v2.contents.bytemaskedarray import ByteMaskedArray
 from awkward._v2.forms.bitmaskedform import BitMaskedForm
@@ -252,7 +251,7 @@ class BitMaskedArray(Content):
         if where < 0:
             where += self.length
         if not (0 <= where < self.length) and self._nplike.known_shape:
-            raise nested_indexerror(self, where)
+            raise ak._v2._util.indexerror(self, where)
         if self._lsb_order:
             bit = bool(self._mask[where // 8] & (1 << (where % 8)))
         else:
