+++ src/awkward/_v2/contents/listarray.py
@@ -4,7 +4,6 @@ import copy
 
 import awkward as ak
 from awkward._v2.index import Index
-from awkward._v2._slicing import nested_indexerror
 from awkward._v2.contents.content import Content
 from awkward._v2.contents.listoffsetarray import ListOffsetArray
 from awkward._v2.forms.listform import ListForm
@@ -166,7 +165,7 @@ class ListArray(Content):
         if where < 0:
             where += self.length
         if not (0 <= where < self.length) and self._nplike.known_shape:
-            raise nested_indexerror(self, where)
+            raise ak._v2._util.indexerror(self, where)
         start, stop = self._starts[where], self._stops[where]
         return self._content._getitem_range(slice(start, stop))
 
@@ -212,7 +211,7 @@ class ListArray(Content):
             nextstarts = self._starts[carry.data]
             nextstops = self._stops[: self._starts.length][carry.data]
         except IndexError as err:
-            raise nested_indexerror(self, carry.data, str(err))
+            raise ak._v2._util.indexerror(self, carry.data, str(err))
 
         return ListArray(
             nextstarts,
@@ -253,7 +252,7 @@ class ListArray(Content):
         slicestarts = slicestarts._to_nplike(self.nplike)
         slicestops = slicestops._to_nplike(self.nplike)
         if slicestarts.length != self.length and self._nplike.known_shape:
-            raise nested_indexerror(
+            raise ak._v2._util.indexerror(
                 self,
                 ak._v2.contents.ListArray(
                     slicestarts, slicestops, slicecontent, None, None, self._nplike
@@ -376,7 +375,7 @@ class ListArray(Content):
             slicecontent, ak._v2.contents.indexedoptionarray.IndexedOptionArray
         ):
             if self._starts.length < slicestarts.length and self._nplike.known_shape:
-                raise nested_indexerror(
+                raise ak._v2._util.indexerror(
                     self,
                     ak._v2.contents.ListArray(
                         slicestarts, slicestops, slicecontent, None, None, self._nplike
@@ -766,7 +765,7 @@ class ListArray(Content):
             headlength = head.length
             head = head._to_nplike(self.nplike)
             if advanced is not None:
-                raise nested_indexerror(
+                raise ak._v2._util.indexerror(
                     self,
                     head,
                     "cannot mix jagged slice with NumPy-style advanced indexing",
