+++ src/awkward/_v2/forms/emptyform.py
@@ -1,7 +1,6 @@
 # BSD 3-Clause License; see https://github.com/scikit-hep/awkward-1.0/blob/main/LICENSE
 
 import awkward as ak
-from awkward._v2.contents.content import nested_indexerror
 from awkward._v2.forms.form import Form, _parameters_equal
 
 
@@ -53,10 +52,10 @@ class EmptyForm(Form):
         )
 
     def _getitem_field(self, where, only_fields=()):
-        raise nested_indexerror(self, where, "not an array of records")
+        raise ak._v2._util.indexerror(self, where, "not an array of records")
 
     def _getitem_fields(self, where, only_fields=()):
-        raise nested_indexerror(self, where, "not an array of records")
+        raise ak._v2._util.indexerror(self, where, "not an array of records")
 
     def _carry(self, allow_lazy):
         return EmptyForm(
