+++ src/awkward/_v2/forms/numpyform.py
@@ -2,7 +2,6 @@
 
 from collections.abc import Iterable
 
-from awkward._v2.contents.content import nested_indexerror
 from awkward._v2.forms.form import Form, _parameters_equal
 
 import awkward as ak
@@ -157,10 +156,10 @@ class NumpyForm(Form):
         )
 
     def _getitem_field(self, where, only_fields=()):
-        raise nested_indexerror(self, where, "not an array of records")
+        raise ak._v2._util.indexerror(self, where, "not an array of records")
 
     def _getitem_fields(self, where, only_fields=()):
-        raise nested_indexerror(self, where, "not an array of records")
+        raise ak._v2._util.indexerror(self, where, "not an array of records")
 
     def _carry(self, allow_lazy):
         return NumpyForm(
